+++ formio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 9.4
+
+- Portal: Fix `/data` URL-path to fetch (Odoo) model records, which required to change the Form URL in (embed) iframe:\
+  - from: `/formio/form/<string:uuid>?portal=true`\
+  - to: `/formio/portal/form/<string:uuid>`
+- Public: Implemented feature `/data` URL-path, to fetch (Odoo) model records.\
+  This configuration requires additional model read-access for the public user, on the model to fetch the records from.
+
 ### 9.3
 
 - Improve the formio.js library Downloader and Installer tool.\